Add preview frames when mousing over a tab
There's some logic to only show the preview if they hover over for a
certain amount of time.  But once they have a preview open it allows
other previews to be opened without that delay.

There's room for improvement in the future to not show the active frame
in between previews when they are in this mode.

This allows us to avoid taking / storing a screenshot of the tab.

Auditors: @diracdeltas