Fix unpinning sites
There was a problem that was made very apparent when https everywhere
was added.  The src is stored in sites for pinned sites and not the
current location.  But for the current window's tab when a tab is first
created it's still the location because we don't want to reload the
frame.

Auditors: @diracdeltas