Closing frame should clear preview frame key. Fixes #182
This does have a minor side effect that a preview won't appear
after closing a tab until you mouse onto a new tab but I think that's
better than what the current state is.

Conditionally clearing the previewFrameKey can lead to a lot more
complexity depending on if the preview is before or after the closing
frame. So clearing is simplest / less complex.

Auditors: @diracdeltas