Add bookmark folders step 1
This adds the ability to add bookmark folders.
It might seem strange that they appear "sites" of
app state but they need to be re-ordered along with the sites so it has
to be that way.

I think eventually by the way we'll split sites into
bookmarks (containing bookmark tags and bookmark folder tags) and pins.
Currently there's a limitation where reordering bookmarks and pins
independently isn't possible.

This is not fully functional bookmark folders
yet but it does allow you to create folders and re-order them along with
bookmarks.  Currently you can't place items in the folders but I'll do
that next.

Auditors: @diracdeltas