Add nestable bookmarks / folders
This allows you to right click and edit a folder and bookmark and put it in a different folder.

After coding this I realized that context menus aren't ideal for viewing
folders because you can't drag them or right click on them once in a menu.  Safari does have this same limitation but it's pretty lame.  I'll refactor this after to be html instead of a context menu.

For now I'm switching gears to add folder support w/ drag and drop into
the bookmark manager.

Auditors: @diracdeltas