Use higher dpi image for about dialog. Fix #833
As discussed this will later go into an about: page.  Using the higher
dpi image always is fine for now because it gets resized anyway.  It is
still a little squashed but temporary and it's as it was.  I think the
icon was replaced at some point with 3 dpi variations and the old filename used to have the high dpi icon inside of it.

Auditors: @aekeus