Convert to HTML based hamburger menu
This will allow us to have more advanced menu items such as z multi zoom click area item.  It also opens up the ability to make the code testable.  One downside is that accelerators are no longer shown on the hamburger menu, but I think we can live with that for now.

Auditors: @aekeus