Change URL only once navigated
This cleans up a lot of page loading handling. In particular we only set the location when a page is navigaed. This fixes when you start the browser you get a save download prompt. It should also fix some trouble with us sometimes clearing the page title

The hacks for mailto: are no longer needed now.

Fix #1219

Auditors: @diracdeltas