Add downloadInProgressUnknownTotal to download.properties
Copied the text from each locale's downloadInProgress property and
replaced the progress percent with a U+2026 (Horizontal Ellipsis)
character. Note that encoding of ellipsis is not consistent across
property files in the project: some use three dots, some use the
single unicode char, with a prevalence of three dots. Despite that,
I opted for the unicode char as the space is tight in the downloads
popup window. In Wikipedia (https://en.wikipedia.org/wiki/Ellipsis#In_different_languages),
I read that in Japanese the ellipsis is actually six dots, but I saw
that in other properties the localizer had used only three so I stuck
to that. Tested only in U.S. English locale in OS X.

Fix #2351