Fix PDFJS MIME type detection
This reverts the old workaround for pdfjs returning status code 0 errors,
which was causing PDF MIME type detection to stop working in cases
where the URL is loaded directly instead of by navigation. A webview
reload seems to always get rid of the PDF status code 0 error.

This includes a merge of PDFJS latest master, but the fix commit is only
https://github.com/diracdeltas/pdf.js/commit/6c185727140372c2541fa45216445d95ed9f9c26

Fix #2633
Fix #2654
Fixes the session tab part of #2568

Auditors: @bbondy