Cache current window
ipcRenderer does sync IPC on each access to getCurrentWindow() this leads to about 30% of the time when typing in the URL bar being used there (the highest function). After this change it goes to 4% for ipcRenderer when typing in the URLbar.

This helps #2736, but I'll do some more investigating for more optimizations

Auditors: @bridiver