URLbar suggestions always 1 char behind
Fix #3643

Auditors: @bridiver

Note the bug was introduced here:
https://github.com/brave/browser-laptop/commit/f83a22ac73f06f2de624f12e43cef1bc28e84bcd
and didn't have an issue, we probably would have tested and fixed process wise if we had an issue and milestone for it.

Test Plan:
- Create at least a couple bookmarks, it's fine if you already have
  some.
- Type `tw` in the urlbar`

Actual results: it will show the bookmarks as long as there
  is a `t` in the title or URL before the fix.

Expected results: It should now not show the bookmarks that don't have
`tw` in them.