Fix open all tabs in about:bookmarks on toolbar
Fix #3614

Auditors: @aekeus

Since this can cause the browser to
hang with a ton of tabs, when it's a lot of tabs I only load the first
30 as loaded tabs.

Test Plan:
Go into about:bookmarks and open in all tabs on the toolbar.
Make sure it works.
Then do the same for some
other folder, make sure it works.  Then test with > 30 items and make
sure only the first 30 show up loaded