Brave doesn't remember last position and size if you set it to start with homepage
When closed Brave does not retain last window position and size as it should be when re-launched.

This happens only if in Settings -> General you set the value for Brave starts with to My homepage. If you select Session and tabs from last time it correctly remembers the position and the size of the window.

Fixes ##3754

Auditors
@bbondy

Test Plan:
1. In settings -> general set Brave starts with to My homepage
2. Launch Brave, move the window around and close
3. Re-launch Brave and compare the last window position