Show top panel in fullscreen on mouse over
Currently when you are in fullscreen mode (F11)
There is no way to see all opened tabs or to type new address
in the url bar. They only way to do it is to switch off fullscreen
and type new address/go to other tab and then switch back to
fullscreen.

This PR makes the top bar behave like the top bar in Chrome and FF.
When you hover with the mouse at the top of the page while in fullscreen
it will appear. Same will happen if you focus the URL bar while in
fullscreen mode (CTRL + L).

Auditors:
@bbondy

Test Plan:
1. Go in fullscreen (F11)
2. Hover whit mouse the top of the page
3. Focus the url bar (CTRL + L)