Fix findbar lag
This changes the findbar from a controlled React element to a
mostly-uncontrolled one, so the characters show up in the input
field before React state is propagated. When the frame is changed
or the findbar component is created for the first time, the input
is restored from saved state.

Auditors: @bbondy

Test Plan:
1. open some tabs
2. cmd+f to find a string in one of the tabs
3. switch to a different tab, cmd+f for a different string
4. switch back to the first tab - the original input should still be in the findbar