Insert virtual history items for sites like www.google.ca
When a site is visted for a deep linked page or with parameters
the history item is stored with the complete parameters and links.
Often after a user visits the site a number of times, the expectation
shifts such that the simple domain should be the first suggestion
in the url bar suggestion list. The canonical example is when a
user searches Google a number of times via the url bar. After three
searches www.google.com will be added as a virtual history item in the
search results.

Auditors: @bbondy, @bradrichter

Test Plan:

  1. Clear history
  2. End three searchs from the url bar without accessing www.google.com
  3. Open a new tab and type goog - ensure a history item for www.google.com is displayed

Fixes: #5067