Fix test failure for notification prompt
There is a race condition with titlemode caused by navigating to about:newtab where it remains in urlbar mode.  But the purpose of this test isn't to handle that, and navigating to about:newtab is not commonly done by users. So just adjusting the test to avoid that

Auditors: @diracdeltas