Improve clarity of the Ubuntu install instructions
What
===
Change the Ubuntu AMD64 installation instructions to show the commands required to append the sources.list entry and complete the full installation, rather than only show the entry that needs to be made.

Why
===
The instructions are not clear that the deb entry needs to be added to the sources.list and the way it is included makes it look like a command the user needs to execute. The instructions are also missing the `apt update` and `apt install brave` that are also necessary.