Site entries without a folderId should not be considered bookmark folders
This code has gone through a lot of changes so it could be an old bug
which doesn't exist anymore.

This was causing a crash due to infinite call stack when creating the
bookmark menu.

Fix #7222

Auditors: @darkdh

I don't know why that entry exists but I couldn't find a reason within an acceptable period of time so I'm moving on