Add the same underscore.string dep as ledger-publisher
Auditors: @bridiver

Fix #7911

It turns out we did have an underscore-string dep top level from argparse, but it was old, and it had a different directory hierarchy.
It had trim under lib, and package.json specified that as the main entry
point, but our electron resolver doesn't check that.

The underscore.string that ledger-parser uses is newer and has the deps directly under underscore.string without a lib folder.
By adding the dep top level, we can ensure we use the later version
which doesn't have an extra directory that our resolver doesn't
understand.