Upgrade to electron-builder 17.1.1
Auditors: @bsclifton, @aekeus

Electron-builder has change significantly, but we've been stuck on a
really old version 2.3.1 because it changed a lot and we didn't keep up
with the changes.

This gets us on the latest 17.1.1 which has support for NOT
rebuilding native modules, and also has things we want in the future
like Ubuntu snaps.

electron-builder added muon support here:
electron-userland/electron-builder#1404 (comment)

We currently only use electron-builder for macOS. But it's really
powerful and we should probably use it everywhere. This task is only to
get our mac back on their tip release.