Stop using activeFrameKey for preview tabs too
This doesn't fix a particular bug that I know how to reproduce, but we
shouldn't be using this anymore.  I do randomly sometimes notice
intermittent problems with preview sticking, and I think it might be
fixed by this.

We do have some test coverage for preview frames but I don't know how to
reproduce the rarely intermittent problem so I can't write a test for this one.