Command+L when in url bar with URL bar suggestion should select full URL
There was also a lot of other fixes which moves things out of the
component and into the store.

Note that there are no state functions in use yet for the window
renderer so I didn't create them yet as part of this task and just
used immutableJS directly.

I also fixed some intermittents with this fix.

Fix #9914

Auditors: @bsclifton

Test Plan:
1. Fully test url bar and suggestions thoroughly
2. Type in some chars that lead to a autocomplete selection, press
command+L and make sure it selects all text.