Fixes dragging into/from Other Bookmarks
Fixes #10157

Bookmarks could not be moved to the Other
Bookmarks folder, where archived bookmarks
reside. The reason surrounded two issues, as
follows:

The first issue was in the renderer process,
which did not correctly set the parentFolderId
on  the keys of bookmarks in child folders. e.g.
Bookmarks Toolbar > Test Folder > Bookmark.

The second issue was in the main process, which
was not updating the `parentFolderId` property
of the old bookmark before setting it anew.

I’ve set two placeholder tests, skipped for now,
as I don’t want to hold up value for the user
while I find my way around what tests already
exist.