Stop losing pinned tabs in a window when adding or removing another pinned tab
Fix #3760, fix #9635, fix #10037, fix #10510, and possibly addresses #10117

The previous code resulted in tabs being marked as no longer required to be pinned, and closed prematurely. This mostly occured after a pinned-tab reordering - afterwards, when pinning or unpinning a tab, all previously pinned tabs with a new order would be closed. Sometimes this appeared like a 'crash' when the window closed because all the tabs had been (undesirably) closed.