Sort simple URL suggestions first again
Fix #11057

Since this has to do with muon url parsing libraries it needed a test inside the muon test as well.
But muon tests are only in master, so I included that test only in
master.

This was introduced when we switched URL parsing libraries because now parsed URL attributes are not null but instead an empty string.