Use addBookmark instead of addSite for bookmarks
Fix #11033

We currently don't have DND under automated so no extra test coverage.

The js/dnd change is if you drag a bookmark from Chrome to Brave it was erroring with a JS error before instead of creating the bookmark.  Now it works.

This is a subset of a larger change that we needed for 0.18.x and 0.19.x
only.

Those changesets are:
582224d005cb107d3e5e31d62b3e9e3e869e82ff
48fee97f42c6841cc3b8b9bd055ea5af11044dec

Auditors: @cezaraugusto