Support drag and drop urlbar icon onto bookmark folders
Fix #11180

This mostly just reverts a subset of:
https://github.com/brave/browser-laptop/commit/d82f17177fbf9effadaedb27e65328344830c7ed

This is only for 0.19.x and 0.20.x

Master can get a better fix so tests and refactoring will be done in:
https://github.com/brave/browser-laptop/issues/11226

Since the tests would not be valid past 0.20.x I'm not covering any new
0.19.x and 0.20.x only tests here.  Also we don't have drag and drop
under tets at all.  In master with the refactor since we have bookmark
toolbar width in state we'll be able to do everything in the reducer and so also be able
to add tests.

Auditors: @bsclifton