Tab previewing: tab gets visual state, content gets gradient opacity mask. Also tighten up some tab styles / spacing as discussed with design. Fix favicon getting cut off at small tab sizes with new spacing.
Tab preview and hover timings which feel more natural for exploring open tabs - fast in and slow out
Tab preview fades out active tab, for focus

Cross-fade of frame content was considered, but removed due to performance / low animation framerate.
This can be fixed and re-implemented with single-webview, which will require a
different css architecture for this potential feature anyway.