Tab title now supports nice subpixel antialiasing
Don't use background mask text, and render on the CPU for better looking text.
Make sure tab is not rendered on GPU when it's not doing anything by not specifying will-change css properties, so text is rendered with subpixel antialising.

Smoother visual transition between selected tabs
Ensures that the gradient used for hiding the title is not visible in the tab's previous colors when a tab switches between active and inactive.
Still allows for transitioning the color of the tab title text.

Tab title text no longer needs different weights for different platforms - rendering is consistent