Prevent null key in historySites state
This was causing an error to be thrown for every subsequent action, and no more state updates to be sent to windows, resulting in multiple issues.
It was also causing an error dump to be generated upon every action, thus halting the browser at those times, including any audio/video (and creating stutters).

Fix #13079 (main test case with STR)
Fix #12333
Fix #12828
Fix #13157
Address #13087 but this is a broad issue - this change will fix a hard issue, whereas that refers to general slowness.