Workaround for aphrodite causing initial css values to be transitioned when they should not be.
Aphrodite does not inject style rules to DOM until some time after they are first used. Unfortunately that creates elements with 0 css rules with rules that are then injected after the elements are added to the DOM, causing any transitions to fire on the initial values, which normally would not be the case. See https://codepen.io/petemill/pen/rdeqqv for demo.