Do not get tab themeColor until window has width
Because a Tab's are no longer immediately attached to a `<webview>`, they do not receive a proper default window size until such a time that it is. Because the DOM window width and height is 0px, we cannot choose elements at certain pixel coordinates in order to find the background color and report as the page's theme color.
Instead, if the page size is 0x0 (which is invalid), we wait for the first resize event.