Ensure tabs are not created in hidden window and correct window is focused
Fix #13859
Fix #13860
Fix #13865

Makes sure tabs are created in a shown window, or creates a new window, that the window is focused, and if the shown window is minimized, then that window is restored.
Fixes for the following scenarios...:
    - There are no windows open
    - There are only minimized windows open
...For the following actions:
    - Brave is default browser and link clicked in another application
    - macOS dock icon is clicked
    - 'Brave' menu actions such as 'New Tab' or 'About Brave'

Test plan:
Browser must be set as default. If testing from source, the best way is to create a build `CHANNEL=dev npm run build-package`, run that, and set it as default.
    - Click URL link in external application
      - Ensure that tab opens in an active window, and the window is focused with the following scenarios:
        - 1 window open
        - 2+ windows open
        - 0 windows open
        - 1 minimized window open
        - 1 minimized window open and 1 non-minimized window open
        - Application exited
      - macOS: click dock icon, ensure window is focused or created:
        - 1 window open
        - 2+ windows open
        - 0 windows open
        - 1 minimized window open
        - 1 minimized window open and 1 non-minimized

Even though the reported issues were macOS only, this test plan should be carried out on other platforms to ensure no regressions.