Hide bookmark bar when removing last item
Fix #14047
The bookmark bar has always remained 'visible' when removing the last item. It just happened to be reduced to a few pixels tall. This was a defect as it affected the vertical rhythm of the UI. With #13758, the toolbar always remains the set height. With this change, the toolbar is hidden automatically when the last item is removed, just like it already automatically shows when the first item is added.

Test Plan:
Automated
- `npm run unittest -- --grep 'bookmarksReducer'`

Manual
- Add several bookmark items, including folders
- Remove some items
  - observe the toolbar is still shown
- Move some items
  - observe the toolbar is still shown
- Delete all items
  - observe the toolbar is not shown anymore
- Recreate items, and use 'move' to put all items in a folder that is not on the bookmark bar
  - observe the toolbar is not shown anymore