When no frame index is provided for a new tab, insert at the end of the frame list
Fix #14101
Whilst we do not display a tab until it has received tab-inserted-at and therefore has an index, we do switch the active Tab Page to the index of the new tab (if it is active), which would be the index of the tab page at tab index 0.

A better fix would be to not add the frame to the window state until tab-inserted-at has been received. This could also help performance, especially if batched.