Rework Tor daemon controller event emitters.
- 'end' means end of input only
- 'error' means an error happened but no state change
- 'close' means underlying resource is now closed

Make sure TorDaemon emits 'exit' event when control closes.  Confirm
in automatic tests that

(a) the TorDaemon emits 'exit', _and_
(b) the tor process exits.

This increases the total amount of code by about 70 lines, but the
logic in each of the event handlers is smaller because their use more
precisely reflects their fine-grained intent in the nodejs APIs, and
many of the new lines are in comments on new short internal
subroutines.