Set frame index immediately when tab inserted to window tab strip.
Fix #14281
A recent change in muon (between 6.0.9 and 6.0.11) has meant that certain tabs (usually with opener tabs) have an invalid index (-1) before they are added to the window tab strip. In such a case, we add the frame at the last index in the frame state. When the tab received a valid window tab strip, the window was notified and the tab UI shown. However, the index change event was notified separately, causing a delay, and a flash of the tab at the last index of the tab strip, followed by a move to the correct index.
Instead, with this change, we set the frame index immediately, as soon as we set the property that informs the tab has been inserted to the window's tab strip.

Validates input just in case we receive an invalid index (e.g. -1). However, there is no known scenario where this could happen at the moment.