Change behaivor of tools/upload.py
Right now multiple builds upload to the same draft, before I wrote
this assuming tools/upload.py was called once. This changes the
behaivor of --force to remove individual files in a draft not the
draft itself. When an existing draft is found it will now instead
merge the files with that draft.