Move tabs toolbar mouseleave handler to parent element which does not use 'contents' value for display style property
Since a `display: contents` container will essentially be the width of its children, and the child was being removed on tab close, the width of the parent element was narrowing so that the mouse was no longer on top, but the mouseleave event did not fire. Moving this handler to a parent element fixes it. It encompasses the pinned tabs but this should not be an issue as the intention of the feature is to relate to the tabs bar as a whole.

Fix #14477