Move control port and control cookie out of the way before reading.
This is an attempt to avoid a race condition in launch where we
inadvertently use the previous run's control port.  There's already
some logic to avoid stale control cookies, but I guess it wasn't
enough here.  Haven't fully reconstructed how this happens but
perhaps this will fix it.

Auditors: @diracdeltas

Test Plan:
1. Use Windows.
2. Open Brave.
3. Enable Tor in private tabs.
4. Confirm private tabs with Tor work.
5. Restart brave.
6. Confirm private tabs with Tor still work.

Previously, on restarting Brave in (5), we would attempt to connect
to the old control port number tor was listening to after (2), as
confirmed by matching up tor.log and console output.