Use host 127.0.0.1 explicitly, instead of defaulting to localhost.
May fix Brave's failure to connect to Tor on some networks, with the
following symptom printed to the console:

    tor: control socket error: Error: getaddrinfo ENOTFOUND localhost undefined:55219
    tor: control socket error: Error: getaddrinfo ENOTFOUND localhost undefined:55219
    tor: control socket closed early

Auditors: @diracdeltas

Test Plan:
Sneak into David's house and try running Brave on his home network.
If he has reconfigured it, invent a time machine to revert it to the
previous state where it didn't work.