Replace missing variable for context menu max-height
Fix #14606
Fixes context menu overflow scroll, returning it to previous behavior. Note that previous behavior is flawed as it only calculates an approximation of available space. It does not cover dynamic 'navbar' height such as menu bar or notifications presence. It could be much more gracefully achieved via `bottom: 0` positioning and would look better by creating a container element which has `max-height: 100%` on it so that the inside of the menu box scrolled rather than the border scrolling away wierdly.