Assert that we've closed only on the next event loop iteration.
The event callbacks may not fire immediately.  We may have to wait
for the readable or writable to notice that they are closed.

It may be that process.nextTick is enough for this, but setImmediate
shouldn't hurt and gives us more time for I/O to happen.

fix #14360

Auditors: @diracdeltas @bsclifton

Test Plan: