Make TorDaemon.kill() idempotent.
This, and TorControl.destroy(), are worth making idempotent because
errors may happen synchronously _or_ asynchronously, so there are two
levels at which a caller might want to clean up.  For example,
authentication failure is reported as a 515 error from tor, _and_ the
tor daemon will close the control connection leading to an
asynchronous error.