On control channel error, continue watching after we report error.
To do this, we separate step of killing the daemon and control channel
from the step of ceasing to watch for updates.  On errors, we want to
kill the daemon and control channel, but continue watching for updates
in case the daemon restarts.

Possible snag: we do not rate-limit failures if they are recursive.
Not currently an issue but we should take care of this in the future.